+++ Source/Language/Squirrel.js
@@ -1,6 +1,6 @@
 /*
 ---
-description: Squirrel Language
+description: Squirrel Language http://www.squirrel-lang.org/
 
 license: MIT-style
 
@@ -28,16 +28,28 @@ EJS.Language.squirrel = new Class({
         
         this.patterns = {
             'slashComments': { pattern: this.common.slashComments, alias: 'co1'},
+
             'poundComments': { pattern: this.common.poundComments, alias: 'co1'},
+
             'multiComments': { pattern: this.common.multiComments, alias: 'co2'},
-            'chars':         { pattern: this.common.singleQuotedString, alias: 'st0' },
+
             'strings':       { pattern: this.common.doubleQuotedString, alias: 'st1' },
-            'annotation':    { pattern: /@[\W\w_][\w\d_]+/gm, alias: 'st1' },
-            'numbers':       { pattern: /\b((([0-9]+)?\.)?[0-9_]+([e][-+]?[0-9]+)?|0x[A-F0-9]+|0b[0-1_]+)\b/gim, alias: 'nu0' },
-            'properties':    { pattern: this.common.properties, alias: 'me0' },
+
+            //'annotation':    { pattern: /@[\W\w_][\w\d_]+/gm, alias: 'st1' },
+
+            // int, float, octals, hex
+            'numbers':       { pattern: /\b((([0-9]+)?\.)?[0-9_]+([e][-+]?[0-9]+)?|0x[A-F0-9]+)\b/gim, alias: 'nu0' },
+
+            // chars are handled as numbers
+            'charnumber':    { pattern: this.common.singleQuotedString, alias: 'nu0' },
+
             'brackets':      { pattern: this.common.brackets, alias: 'br0' },
-            'functionCalls': { pattern: this.common.functionCalls, alias: 'de1'},
-            'directives':	 { pattern: /#.*$/gm, alias: 'kw2'}
+
+            'functionCalls': { pattern: this.common.functionCalls, alias: 'me0'},
+
+            'methodCalls':   { pattern: this.common.methodCalls, alias: 'me1'},
+
+            'properties':    { pattern: this.common.properties, alias: 'me1' }
         };
     }
 });